no empty zips

+++ compact.sh
@@ -9,4 +9,8 @@ if [[ -f "$1.tar.gz" ]]; then
     exit 0
 fi
 
+if [[ ! -d histories ]]; then
+    exit 0
+fi
+
 tar -zcvf "$1.tar.gz" histories/ && rm -r histories
